Track all remote reuests sent via AWS SDK
When using the `HM\Platform\get_aws_sdk()`, we add a stats collector callback, so we can send all data of remote requests to AWS services to XRay.

I spend a long time, way way too long, trying to figure out how to wrap the AWS handlers / middleware via the client init params, but I just don't think it's possible to do that. Someone else way more experience with the AWS SDK may be able to do this, but I was not able to make good progress after several hours.

I ultimately had to go with using the Guzzle's `on_stats` callback. This is a lot lower level, we get the HTTP request and the HTTP response, but not hte higher level AWS Command object which would give us "real" access to the AWS Service etc. Turns that I was mostly able to work around that by digging what data we need out of the HTTP request and response though. This is certainly a huge improvement over what we have right now.