Add Query Monitor support
Working on X-Ray locally can be quite tough, as it requires either hooking it up to a live AWS account, or manually printing out data that would be sent to the daemon. This adds a new Query Monitor panel for all the segments sent to AWS X-Ray for inspection. It also adds the Flamegraph for the request in a new panel, so it can be used as a profiling tool too.

Screenshots:

![](https://joehoyle-captured.s3.amazonaws.com/lFfnXjrG.png)

![](https://joehoyle-captured.s3.amazonaws.com/aLESFpXr.png)