Add time stats to main trace including object cache time
Include all time stats for Database, Object Cache and Remote Requests. This makes it easier / possible to generate a pie chart (for example) of script execution time summary.

This also adds the object-cache as it's own service to XRay, so we can get similar visualizations in Xray for object cache like we do db and remote requests.