Handle missing x-amz-target header
In some cases - particularly on local environments `$request_id` can be truthy but empty, or without offset 0 being set, additionally the `x-amz-target` header isn't always set in requests to S3. This update ensures both cases don't cause warnings to be output but rather falls back to reasonable defaults.

A missing `x-amz-target` header will give `amz-unknown-target` as the value if it's not set.