Calculate duration using rounded figure, not raw
In cases where the rounded figure differed by a large amount, this lead
to a large discrepency in the timestamps being calculated.

For example, if there are 1001 frames, we take every second item
($pluck_every_n = 2). However, we calculate the duration as $original *
( 1001 / 1000 ), whereas by halving the number of frames, we actually
need to double the duration.

Using the same variable corrects for this inaccuracy and ensures
consistency across the calculation.